requirements.txt
anyio==4.6.2.post1
APScheduler==3.6.3
cachetools== 4.2.2
certifi==2024.8.30
h11==0.14.0
httpcore== 1.0.6
httpx==0.27.2
idna== 3.10
pip==24.2
python-telegram-bot==13.15
pytz== 2024.2
setuptools== 65.5.1
six==1.16.0
sniffio==1.3.1
tornado==6.1
typing_extensions==4.12.2
tzdata== 2024.2
tzlocal==5.2
wheel==0.38.4
